[refactoring] : interface 제거
