+++ src/main/java/com/example/jolvre/auth/api/AuthController.java
@@ -1,4 +1,11 @@
 package com.example.jolvre.auth.api;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@Slf4j
 public class AuthController {
 }
