+++ src/main/java/com/example/jolvre/chat/service/ChatService.java
@@ -1,4 +1,4 @@
 package com.example.jolvre.chat.service;
 
-public interface ChatService {
+public class ChatService {
 }
