--- src/main/java/com/example/jolvre/chat/service/ChatServiceImpl.java
@@ -1,4 +1,0 @@
-package com.example.jolvre.chat.service;
-
-public class ChatServiceImpl implements ChatService {
-}
