+++ src/main/java/com/example/jolvre/exhibition/service/DiaryPictureService.java
@@ -1,12 +1,27 @@
 package com.example.jolvre.exhibition.service;
 
 import com.example.jolvre.exhibition.entity.DiaryPicture;
+import com.example.jolvre.exhibition.repository.DiaryPictureRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface DiaryPictureService {
+@Service
+@RequiredArgsConstructor
+public class DiaryPictureService {
+    private final DiaryPictureRepository diaryPictureRepository;
 
-    public DiaryPicture upload(DiaryPicture diaryPicture);
 
-    public DiaryPicture update(DiaryPicture diaryPicture);
+    public DiaryPicture upload(DiaryPicture diaryPicture) {
+        return diaryPictureRepository.save(diaryPicture);
+    }
 
-    public void delete(DiaryPicture diaryPicture);
+
+    public DiaryPicture update(DiaryPicture diaryPicture) {
+        return null;
+    }
+
+    
+    public void delete(DiaryPicture diaryPicture) {
+        diaryPictureRepository.delete(diaryPicture);
+    }
 }
