+++ src/main/java/com/example/jolvre/exhibition/service/DiaryService.java
@@ -1,11 +1,24 @@
 package com.example.jolvre.exhibition.service;
 
 import com.example.jolvre.exhibition.entity.Diary;
+import com.example.jolvre.exhibition.repository.DiaryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface DiaryService {
-    public Diary upload(Diary diary);
+@Service
+@RequiredArgsConstructor
+public class DiaryService {
+    private final DiaryRepository diaryRepository;
 
-    public Diary update(Diary diary);
+    public Diary upload(Diary diary) {
+        return diaryRepository.save(diary);
+    }
 
-    public void delete(Diary diary);
+    public Diary update(Diary diary) {
+        return null;
+    }
+
+    public void delete(Diary diary) {
+        diaryRepository.delete(diary);
+    }
 }
