+++ src/main/java/com/example/jolvre/exhibition/service/ExhibitService.java
@@ -1,11 +1,24 @@
 package com.example.jolvre.exhibition.service;
 
 import com.example.jolvre.exhibition.entity.Exhibit;
+import com.example.jolvre.exhibition.repository.ExhibitRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface ExhibitService {
-    public Exhibit upload(Exhibit exhibit);
+@Service
+@RequiredArgsConstructor
+public class ExhibitService {
+    private final ExhibitRepository exhibitRepository;
 
-    public Exhibit update(Exhibit exhibit);
+    public Exhibit upload(Exhibit exhibit) {
+        return exhibitRepository.save(exhibit);
+    }
 
-    public void delete(Exhibit exhibit);
+    public Exhibit update(Exhibit exhibit) {
+        return null;
+    }
+
+    public void delete(Exhibit exhibit) {
+        exhibitRepository.delete(exhibit);
+    }
 }
