+++ src/main/java/com/example/jolvre/post/service/CommentService.java
@@ -1,11 +1,24 @@
 package com.example.jolvre.post.service;
 
 import com.example.jolvre.post.entity.Comment;
+import com.example.jolvre.post.repository.CommentRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface CommentService {
-    public Comment upload(Comment comment);
+@Service
+@RequiredArgsConstructor
+public class CommentService {
+    private final CommentRepository commentRepository;
 
-    public Comment update(Comment comment);
+    public Comment upload(Comment comment) {
+        return commentRepository.save(comment);
+    }
 
-    public void delete(Comment comment);
+    public Comment update(Comment comment) {
+        return null;
+    }
+
+    public void delete(Comment comment) {
+        commentRepository.delete(comment);
+    }
 }
